Add proto.rs to include protocol related.
Move the MessageHeader from `common.rs` to the `proto.rs` file.
Move the `compiled::ttrpc` from `common.rs` to `proto.rs` file and re-export.
