+++ src/sync/channel.rs
@@ -15,10 +15,8 @@
 use nix::sys::socket::*;
 use std::os::unix::io::RawFd;
 
-use crate::common::{MESSAGE_HEADER_LENGTH, MESSAGE_LENGTH_MAX};
 use crate::error::{get_rpc_status, sock_error_msg, Error, Result};
-use crate::proto::Code;
-use crate::MessageHeader;
+use crate::proto::{Code, MessageHeader, MESSAGE_HEADER_LENGTH, MESSAGE_LENGTH_MAX};
 
 fn retryable(e: nix::Error) -> bool {
     use ::nix::Error;
